Fix bug when DELETE ACID block is a DictionaryBlock
Before this commit, Block.getChildren() was used to take apart the
DELETE ACID rowId block.  That is always the wrong thing to do.
Fixed by using ColumnarRow to access the elementa of the ACID rowId block.

This commit also adds a comment warning developers not to use
Block.getChildren() so others don't make the same mistake of
calling the method.